Cython Bindings / Utils (#1348)
* Types only for type annotations

* KeysIterator and ValuesIterator wrappers

* Cython bindings

* Add # noqa F401

* Use public `.load()` intead of `.data`
