+++ aim/storage/types.py
@@ -1,6 +1,6 @@
-from typing import Callable, Dict, List, Tuple, Union, Generic, TypeVar
-from copy import deepcopy
+from typing import Dict, List, Tuple, Union
 
+from aim.storage.utils import BLOB  # noqa F401
 
 NoneType = type(None)
 
@@ -65,60 +65,3 @@ class SafeNone(metaclass=Singleton):
 
 class CustomObjectBase:
     pass
-
-
-BLOBLoader = Callable[[], AimObjectPrimitive]
-T = TypeVar('T')
-
-
-class BLOB(Generic[T]):
-    def __init__(
-        self,
-        data: T = None,
-        loader_fn: 'BLOBLoader' = None
-    ):
-        self.data: T = data
-        self.loader_fn = loader_fn
-
-    def __bytes__(self):
-        return bytes(self.load())
-
-    def load(self) -> T:
-        if self.data is None:
-            assert self.loader_fn is not None
-            self.data = self.loader_fn()
-            self.loader_fn = None
-        return self.data
-
-    def __deepcopy__(self, memo):
-        data = self.load()
-        instance = self.__class__(data=deepcopy(data, memo=memo))
-        memo[id(self)] = instance
-        return instance
-
-    def transform(
-        self,
-        transform_fn: Callable[[AimObjectPrimitive], T]
-    ) -> 'BLOB':
-        if self.data is not None:
-            return self.__class__(transform_fn(self.data))
-
-        def loader():
-            return transform_fn(self.load())
-
-        return self.__class__(loader_fn=loader)
-
-
-__all__ = [
-    'NoneType',
-    'AimObjectKey',
-    'AimObjectPath',
-    'AimObjectPrimitive',
-    'AimObjectArray',
-    'AimObjectDict',
-    'AimObject',
-    'CustomObjectBase',
-    'SafeNone',
-    'BLOB',
-    'BLOBLoader',
-]
