+++ aim/storage/utils.pxd
@@ -1,0 +1,20 @@
+# distutils: language = c++
+# cython: language_level = 3
+
+cdef class ArrayFlagType:
+    pass
+
+cdef class ObjectFlagType:
+    pass
+
+cdef class CustomObjectFlagType:
+    cdef:
+        public str aim_name
+
+cdef class BLOB:
+    cdef:
+        object data
+        object loader_fn
+    cpdef bytes load(self)
+    # TODO closures inside cython functions are not supported yet
+    # cpdef object transform(self, object transform_fn)
