+++ aim/storage/utils.py
@@ -1,3 +1,30 @@
+from copy import deepcopy
+
+
+class KeysIterator:
+    def __init__(self, items_iterator):
+        self.it = items_iterator
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        key, _ = next(self.it)
+        return key
+
+
+class ValuesIterator:
+    def __init__(self, items_iterator):
+        self.it = items_iterator
+
+    def __iter__(self):
+        return self
+
+    def __next__(self):
+        _, value = next(self.it)
+        return value
+
+
 class ArrayFlagType:
     def __repr__(self):
         return "<ArrayFlag>"
@@ -34,8 +61,36 @@ class CustomObjectFlagType:
         return f"<CustomObjectFlag type={self.aim_name}>"
 
 
-__all__ = [
-    'ArrayFlag',
-    'ObjectFlag',
-    'CustomObjectFlagType',
-]
+class BLOB:
+    def __init__(
+        self,
+        data=None,
+        loader_fn=None
+    ):
+        self.data = data
+        self.loader_fn = loader_fn
+
+    def __bytes__(self):
+        return bytes(self.load())
+
+    def load(self):
+        if self.data is None:
+            assert self.loader_fn is not None
+            self.data = self.loader_fn()
+            self.loader_fn = None
+        return self.data
+
+    def __deepcopy__(self, memo):
+        data = self.load()
+        instance = self.__class__(data=deepcopy(data, memo=memo))
+        memo[id(self)] = instance
+        return instance
+
+    def transform(self, transform_fn):
+        if self.data is not None:
+            return self.__class__(transform_fn(self.data))
+
+        def loader():
+            return transform_fn(self.load())
+
+        return self.__class__(loader_fn=loader)
